planner: don't access current_dive and dc_number globals
Pass down these data from the caller in an attempt to slowly
get rid of global accesses.

Moreover, let the caller copy onto the displayed dive. The
old code was logically unsound: it displayed the displayed_dive
in the profile widget and then let the planner copy over
the dive. And it displayed the first dc, however the replanned
dive might be a different dc, if dives were merged.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>